+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Add the `SearchPage` GraphQL queries.
+- Data layer intergration in the `ProductPage`.
 
 ## [0.3.0] - 2018-6-25
 ### Added