+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add the `DepartmentPage` component.
+
+### Fixed
+- Fix loading logic inside the `SearchQueryContainer`.
 
 ## [0.5.0] - 2018-7-5
 ### Added