+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Graphql product query to support benefits retrievement.
+
 ### Added
 - Add a generic component (`SearchQueryContainer`) to be reused into the different pages that uses search and facets queries.